Create Type of array.py
Given an array having no duplicates, which can be of 4 types
1.  Ascending
2.  Descending
3.  Descending Rotated
4.  Ascending Rotated
Your task is to print which type of array it is and the maximum element of that array.

Examples:

Input :  A[] = { 2, 1, 5, 4, 3}
Output : Descending rotated with maximum element 5

Input :  A[] = { 3, 4, 5, 1, 2}
Output : Ascending rotated with maximum element 5

Input:
The first line of input contains an integer T denoting the no test cases. Then T test caes follow. The first line of input contains an integer n denoting the size of the array. Then in the next line are N space separated values of the array.

Output:
For each test case print the largest element in the array and and integer x denoting the type of the array.

Constraints:
1<=T<=100
4<=n<=100
1<=A[]<=10000

Example:
Input:
2
5
2 1 5 4 3
5
3 4 5 1 2

Output:
5 3
5 4